requirements.txt
mysqlclient ~= 2.1.0
SQLAlchemy ~= 1.4.31
mysql-connector-python
pandas